Update section header for listed devices
The term "supported" could be misconstrued to mean that there is some kind of guarantee, when in reality the information is just provided by individuals who have no connection to the manufacturers or inside knowledge about differences between hardware and firmware versions and very little access to either for testing. Probably this should have "... based on tests with single machines" added, but maybe that's a bit long for a title?